[PC-361] fix: 프로필 기본정보 업데이트 분리
